requirements.txt
pdfplumber
sentence-transformers
faiss-cpu
anthropic
streamlit
spacy
pandas
numpy
python-dotenv
